+++ xenon-common/src/main/java/com/vmware/xenon/services/common/QueryTask.java
@@ -204,9 +204,6 @@ public class QueryTask extends ServiceDocument {
              * Query pages will be deleted immediately after first access instead of at query task
              * expiration time, and any index service resources backing the query will be deleted
              * after the last page is accessed.
-             *
-             * Note this option disables sharing of index service resources with other queries and
-             * may increase overall resource usage.
              */
             SINGLE_USE,